+++ JsMerge.Core/Main.cs
@@ -1,10 +1,10 @@
 ﻿using Newtonsoft.Json;
-using System.Text;
 
 namespace JsMerge.Core
 {
 	public static class Main
 	{
+		public const string VERSION = "1.2.0";
 		#region [Sub systems]
 		/// <summary>
 		/// A subsystem for handling query strings